// src/interfaces/countries.interface.ts
/**
 * Copyright 2021 Crown Copyright (Single Trade Window)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/* eslint-disable camelcase */

export interface Attributes {
  id: string;
  description: string;
  geographical_area_id: string;
}

export interface ChildrenGeographicalAreas {
  data?: (null)[] | null;
}

export interface Relationships {
  children_geographical_areas: ChildrenGeographicalAreas;
}

export interface Countries {
  id: string;
  type: string;
  attributes: Attributes;
  relationships: Relationships;
}

export interface CountriesDropdown {
  value: string;
  text: string;
  selected: string;
}

export default Countries;